refactor(analysis): co-locate uncertainty propagation with its fit
Move each propagation helper into the file that owns the corresponding fit
(_leppr.py, _lambda.py, error_budget/) and drop _propagation.py, so the maths
and its uncertainty propagation live together.

Addresses review feedback on #274. Relates to #215.
